@@ -14,13 +14,13 @@
         }
     ],
     "require": {
-        "php": ">=7.3",
-        "cakephp/cakephp": "^4.2"
+        "php": ">=8.1",
+        "cakephp/cakephp": "5.x-dev"
     },
     "require-dev": {
         "phpunit/phpunit": "^9.5.0",
         "cakephp/cakephp-codesniffer": "^4.5",
-        "dereuromark/cakephp-ide-helper": "^1.0",
+        "dereuromark/cakephp-ide-helper": "dev-cake5",
         "phpstan/phpstan": "^1.0.0"
     },
     "autoload": {
@@ -36,6 +36,7 @@
         }
     },
     "scripts": {
+        "test": "phpunit",
         "stan": "vendor/bin/phpstan analyze",
         "cs-check": "phpcs --colors --parallel=16 -p src/ tests/",
         "cs-fix": "phpcbf --colors --parallel=16 -p src/ tests/"
@@ -44,5 +45,6 @@
         "allow-plugins": {
             "dealerdirect/phpcodesniffer-composer-installer": true
         }
-    }
+    },
+    "minimum-stability": "dev"
 }
