@@ -30,7 +30,7 @@ class ServicePaginatorTraitTest extends TestCase
      *
      * @var array
      */
-    public $fixtures = [
+    protected array $fixtures = [
         'core.Articles',
     ];
 
